feat: pull out the previewe actions
