+++ src/client/theme-default/slots/PreviewerActionsExtra/index.tsx
@@ -1,0 +1,7 @@
+import React, { type FC } from 'react';
+
+const PreviewerActionsExtra: FC = () => {
+  return <></>;
+};
+
+export default PreviewerActionsExtra;
